refactor: rename folder name
